requirements.txt
transformers
torch
Pillow
accelerate
scikit-image
torchvision